Build tables and apply Cell Key Method on a list

This function constructs multiple tables from microdata and applies the Cell Key Method to each table based on a description data frame specifying the table structure.

Usage

tabulate_and_apply_ckm_list(
  df,
  rk_var = "rkey",
  desc_tab,
  marge_label = "Total",
  prefix = "tab",
  D,
  V,
  js = 0,
  I = NULL,
  J = NULL,
  ...
)

Arguments

df

data.frame. Input microdata

rk_var

Character. Individual key variable (NULL for no keys)

desc_tab

data.frame. Table description with 3 columns (TAB, VAR, HRC) describing the tables to construct

marge_label

Character. Margin label (default: "Total")

prefix

character. Prefix to add to table names provided in desc_tab (default: "tab")

D

integer. Deviation parameter (must be strictly positive)

V

numeric. Noise variance (must be strictly positive)

js

integer. Threshold for sensitive values (default: 0). If js=0, only value 0 will be forbidden

I

integer vector. Original values to consider

J

integer vector. Perturbed values to consider

...

Additional parameters passed to transition matrix creation

Value

A list containing:

tab: list of tables (tibbles) with CKM applied to each

ptab: transition matrix used for calculations

risque: tibble with risk measures for each table

utilite: tibble with utility measures for each table

Details

The desc_tab data frame must have the following structure:

TAB: Table name or number

VAR: Variable name for the table

HRC: Hierarchy name if the variable has hierarchical relationship, NA otherwise

Variables with hierarchical relationships should be listed in decreasing order of hierarchy (from broadest to finest level).

Le data.frame à fournir dans l'argument desc_tab doit être constitué de telle façon que:

1) Le data.frame doit contenir les 3 colonnes suivantes: * TAB: Nom ou numéro de la table * VAR: Nom d'une variable de la table * HRC: Si la variable a une relation hiérarchique avec une autre des variables de la table mentionner un nom pour cette hiérarchie. Sinon, NA

2) Une ligne correspond à une variable d'une table donnée

3) Deux variables (ex: REGION et DEPARTEMENT) d'une même table ayant une relation hiérarchique doivent recevoir la même valeur dans la colonne HRC. Ces variables doivent être renseignées dans l'ordre décroissant de la hiérarchie, cad du niveau le plus large (ex: REGION) au niveau le plus fin (ex: DEPARTEMENT).

Examples

if (FALSE) { # \dontrun{
data("dtest")
set.seed(123)
dtest_avec_cles <- build_individual_keys(dtest)

# Define two tables:
# tab1: DIPLOME * SEXE * AGE
# tab2: DIPLOME * TYPE * REG * DEP, where REG > DEP
desc_tableaux <- data.frame(
  TAB = c(rep(1,3), rep(2,4)),
  VAR = c("DIPLOME", "SEXE", "AGE", "DIPLOME", "TYPE", "REG", "DEP"),
  HRC = c(rep(NA, 5), rep("GEO",2))
)

res_ckm <- tabulate_and_apply_ckm_list(
  df = dtest_avec_cles,
  desc_tab = desc_tableaux,
  marge_label = "Total",
  D = 10, V = 15, js = 4
)
} # }